Implement timer!!

+++ am-kernels/benchmarks/dhrystone/dry.c
@@ -354,7 +354,7 @@
 #include <klib.h>
 #include <klib-macros.h>
 
-static uint32_t uptime_ms() { return io_read(AM_TIMER_UPTIME).us / 1000; }
+static long uptime_ms() { return io_read(AM_TIMER_UPTIME).us / 1000; }
 #define Start_Timer() Begin_Time = uptime_ms()
 #define Stop_Timer()  End_Time   = uptime_ms()
 
@@ -795,6 +795,7 @@ int main ()
     /***************/
 
     Start_Timer();
+		//printf("Begin_Time: %d\n",io_read(AM_TIMER_UPTIME).us);
 
     for (Run_Index = 1; Run_Index <= Number_Of_Runs; ++Run_Index)
     {
@@ -847,7 +848,6 @@ int main ()
     /**************/
 
     Stop_Timer();
-
     User_Time = End_Time - Begin_Time;
 
     Done = true;
